[fix]: 스티커박스 - 나타난 오류 전부 해결

+++ Sticker Box/Item.js
@@ -6,12 +6,13 @@ export class Item {
     this.itemEl = this.generateItem();
   }
 
+  // Item HTML 생성
   generateItemUI() {
     const item = document.createElement('li');
     const itemTitle = document.createElement('div');
     const itemDeleteBtn = document.createElement('button');
 
-    item.id = `item-${Item.itemIndex}`;
+    item.id = `${Item.itemIndex}`;
     item.classList.add('item');
     itemTitle.innerText = `Text ${Item.itemIndex++}`;
     itemDeleteBtn.innerText = '삭제';
@@ -23,37 +24,34 @@ export class Item {
     return item;
   }
 
+  // Item에 이벤트 붙이기
   patchItemEvent(item) {
-    item.addEventListener('click', (e) => {
+    item.addEventListener('mousedown', (e) => {
+      e.stopPropagation();
+
       const tagName = e.target.tagName.toLowerCase();
 
-      if (tagName === 'button') {
-        this.deleteItem(this);
-      } else if (tagName === 'li') {
-        this.moveItem(this);
-      }
+      if (tagName === 'button') this.deleteItem(this);
+      else if (tagName === 'li') this.moveItem(e, this);
     });
 
     return item;
   }
 
   generateItem() {
-    // UI
     const itemUI = this.generateItemUI();
-    const item = this.patchItemEvent(itemUI);
-
-    return item;
+    return this.patchItemEvent(itemUI);
   }
 
   // 아이템 삭제
-  deleteItem(item) {
-    const itemsDiv = item.itemEl.parentElement;
+  deleteItem(targetItem) {
+    const itemsDiv = targetItem.itemEl.parentElement;
     const deleteIdx = this.sticker.items.findIndex(
-      (stickerItem) => stickerItem === item
+      (item) => item === targetItem
     );
 
     this.sticker.items.splice(deleteIdx, 1);
-    item.itemEl.remove();
+    targetItem.itemEl.remove();
 
     itemsDiv.style.height = `${itemsDiv.clientHeight - 50.2}px`;
 
@@ -61,181 +59,198 @@ export class Item {
   }
 
   // 아이템 이동
-  moveItem(item) {
+  moveItem(e, target) {
     let draggingItem = null;
-    let originStickerCount = 0;
+    let isFirstFromSticker = true;
 
-    item.itemEl.addEventListener('mousedown', (e) => {
-      e.stopPropagation();
+    const targetEl = target.itemEl;
 
-      draggingItem = item.itemEl;
+    // 드래그 할 아이템 지정
+    draggingItem = targetEl;
 
-      item.sticker.setTop(item.sticker.stickerEl);
-      item.itemEl.style.cursor = 'grabbing';
+    target.sticker.setTop(target.sticker.stickerEl);
+    targetEl.style.cursor = 'grabbing';
 
-      // 원래 자리 기억
-      const originalY = item.itemEl.style.top;
-      const originalZIndex = item.itemEl.style.zIndex;
+    // 원래 자리 기억
+    const originInfo = {
+      y: targetEl.style.top,
+      zIndex: targetEl.style.zIndex,
+    };
 
-      const itemsDiv = item.itemEl.parentElement;
+    const itemsDiv = targetEl.parentElement;
 
-      const parentRect = itemsDiv.getBoundingClientRect();
-      const itemRect = item.itemEl.getBoundingClientRect();
+    const parentRect = itemsDiv.getBoundingClientRect();
+    const itemRect = targetEl.getBoundingClientRect();
 
-      const itemInnerX = e.clientX - itemRect.left;
-      const itemInnerY = e.clientY - itemRect.top;
+    const itemInnerX = e.clientX - itemRect.left;
+    const itemInnerY = e.clientY - itemRect.top;
 
-      let insertIndex;
-      let toSticker = item.sticker;
-      let isOriginSticker = true;
-      let originIndex = item.sticker.items.findIndex(
-        (i) => i.itemEl.id.split('-')[1] === item.itemEl.id.split('-')[1]
-      );
+    let insertIndex;
+    let toSticker = target.sticker;
+    let isFromSticker = true;
+    let fromStickerIndex = target.sticker.items.findIndex(
+      (i) => i.itemEl.id === targetEl.id
+    );
 
-      const onMouseMove = (moveEvent) => {
-        const x = moveEvent.clientX;
-        const y = moveEvent.clientY;
+    const onMouseMove = (moveEvent) => {
+      const clientX = moveEvent.clientX;
+      const clientY = moveEvent.clientY;
 
-        item.itemEl.style.left = `${x - itemInnerX - parentRect.left}px`;
-        item.itemEl.style.top = `${y - itemInnerY - parentRect.top}px`;
-        item.itemEl.style.zIndex = 1000;
-        item.itemEl.style.cursor = 'grabbing';
+      targetEl.style.left = `${clientX - itemInnerX - parentRect.left}px`;
+      targetEl.style.top = `${clientY - itemInnerY - parentRect.top}px`;
+      targetEl.style.zIndex = 1000;
+      targetEl.style.cursor = 'grabbing';
 
-        const closestItemEl = document.elementsFromPoint(x, y)[1];
+      // 가장 가까운 item 찾기
+      const closestItemEl = document.elementsFromPoint(clientX, clientY)[1];
+
+      if (closestItemEl && closestItemEl.classList.contains('items-div')) {
+        const stickers = target.sticker.stickerBox.stickers;
+
+        for (const sticker of stickers) {
+          if (closestItemEl.parentElement === sticker.stickerEl) {
+            toSticker = sticker;
+            isFromSticker =
+              closestItemEl.parentElement === target.sticker.stickerEl;
+            break;
+          }
+        }
+      } else if (
+        closestItemEl &&
+        closestItemEl.classList.contains('item') &&
+        closestItemEl !== draggingItem
+      ) {
+        const stickers = target.sticker.stickerBox.stickers;
+        let closestItem;
+
+        outer: for (const sticker of stickers) {
+          const items = sticker.items;
+
+          for (let i of items) {
+            if (i.itemEl === closestItemEl) {
+              closestItem = i;
+              toSticker = sticker;
+              isFromSticker =
+                target.sticker.stickerIndex === sticker.stickerIndex;
+              break outer;
+            }
+          }
+        }
+
+        // 옮겨갈 sticker에 공간 마련하기
         const isTop =
           closestItemEl.style.height / 2 >
           document.clientX - closestItemEl.getBoundingClientRect().top;
-        console.log(closestItemEl);
-        if (
-          closestItemEl &&
-          closestItemEl.classList.contains('item') &&
-          closestItemEl !== draggingItem
-        ) {
-          const stickers = item.sticker.stickerBox.stickers;
-          let closestItem;
 
-          outer: for (const sticker of stickers) {
-            const items = sticker.items;
-
-            for (let i of items) {
-              if (i.itemEl === closestItemEl) {
-                closestItem = i;
-                toSticker = sticker;
-                isOriginSticker = item.sticker === sticker;
-
-                break outer;
-              }
-            }
+        // 원래 스티커 내에서의 이동
+        if (isFromSticker) {
+          if (isFirstFromSticker) {
+            isFirstFromSticker = false;
+            toSticker.items.splice(fromStickerIndex, 1);
+            toSticker.renderItem(toSticker.stickerEl.children[3]);
           }
 
-          // 옮겨갈 sticker에 공간 마련하기
-          if (isOriginSticker) {
-            if (originStickerCount++ === 0) {
-              toSticker.items.splice(originIndex, 1);
-              toSticker.renderItem(toSticker.stickerEl.children[3]);
+          insertIndex = toSticker.items.findIndex(
+            (item) =>
+              item !== 'EMPTY' && item.itemEl.id === closestItem.itemEl.id
+          );
+
+          if (toSticker.items[insertIndex] !== 'EMPTY') {
+            if (toSticker.items.includes('EMPTY')) {
+              toSticker.items = toSticker.items.filter(
+                (item) => item !== 'EMPTY'
+              );
             }
 
-            insertIndex = toSticker.items.findIndex(
-              (item) =>
-                item !== 'EMPTY' &&
-                item.itemEl.id.split('-')[1] ===
-                  closestItem.itemEl.id.split('-')[1]
-            );
+            toSticker.items = [
+              ...toSticker.items.slice(0, insertIndex),
+              'EMPTY',
+              ...toSticker.items.slice(insertIndex),
+            ];
 
-            if (toSticker.items[insertIndex] !== 'EMPTY') {
-              if (toSticker.items.includes('EMPTY')) {
-                toSticker.items = toSticker.items.filter(
-                  (item) => item !== 'EMPTY'
-                );
-              }
-
-              toSticker.items = [
-                ...toSticker.items.slice(0, insertIndex),
-                'EMPTY',
-                ...toSticker.items.slice(insertIndex),
-              ];
-
-              toSticker.renderItem(toSticker.stickerEl.children[3]);
-            }
-          } else {
-            insertIndex = toSticker.items.findIndex(
-              (item) =>
-                item !== 'EMPTY' &&
-                item.itemEl.id.split('-')[1] ===
-                  closestItem.itemEl.id.split('-')[1]
-            );
-
-            if (isTop && insertIndex === 1) insertIndex--;
-
-            if (toSticker.items[insertIndex] !== 'EMPTY') {
-              if (toSticker.items.includes('EMPTY')) {
-                toSticker.items = toSticker.items.filter(
-                  (item) => item !== 'EMPTY'
-                );
-              }
-
-              toSticker.items = [
-                ...toSticker.items.slice(0, insertIndex),
-                'EMPTY',
-                ...toSticker.items.slice(insertIndex),
-              ];
-
-              toSticker.renderItem(toSticker.stickerEl.children[3]);
-            }
+            toSticker.renderItem(toSticker.stickerEl.children[3]);
           }
         }
-      };
 
-      const onMouseUp = () => {
-        // 원래 자리로 돌아가기
-        item.itemEl.style.left = '0px';
-        item.itemEl.style.top = originalY;
-        item.itemEl.style.zIndex = originalZIndex;
-        item.itemEl.style.cursor = 'grab';
+        // 다른 스티커로의 이동
+        else {
+          insertIndex = toSticker.items.findIndex(
+            (item) =>
+              item !== 'EMPTY' && item.itemEl.id === closestItem.itemEl.id
+          );
 
-        console.log(toSticker);
+          if (isTop && insertIndex === 1) insertIndex--;
 
-        if (isOriginSticker) {
-          console.log('원래');
-          toSticker.items[insertIndex ?? originIndex] = item;
-          toSticker.renderItem(toSticker.stickerEl.children[3]);
-          originStickerCount = 0;
+          if (toSticker.items[insertIndex] !== 'EMPTY') {
+            if (toSticker.items.includes('EMPTY')) {
+              toSticker.items = toSticker.items.filter(
+                (item) => item !== 'EMPTY'
+              );
+            }
 
-          const stickers = item.sticker.stickerBox.stickers;
+            toSticker.items = [
+              ...toSticker.items.slice(0, insertIndex),
+              'EMPTY',
+              ...toSticker.items.slice(insertIndex),
+            ];
 
-          for (let sticker of stickers) {
-            if (toSticker.stickerIndex === sticker.stickerIndex) continue;
+            toSticker.renderItem(toSticker.stickerEl.children[3]);
+          }
+        }
+      }
+    };
 
-            const emptyIndex = sticker.items.indexOf('EMPTY');
+    const onMouseUp = () => {
+      // 원래 자리로 돌아가기
+      targetEl.style.left = '0px';
+      targetEl.style.top = originInfo.y;
+      targetEl.style.zIndex = originInfo.zIndex;
+      targetEl.style.cursor = 'grab';
+
+      // 원래 스티커 내에서의 이동
+      if (isFromSticker) {
+        toSticker.items[insertIndex ?? fromStickerIndex] = target;
+        toSticker.renderItem(toSticker.stickerEl.children[3]);
+        isFirstFromSticker = true;
+
+        const stickers = target.sticker.stickerBox.stickers;
+        for (let sticker of stickers) {
+          if (toSticker.stickerIndex === sticker.stickerIndex) continue;
+
+          const emptyIndex = sticker.items.indexOf('EMPTY');
+          if (emptyIndex !== -1) {
             sticker.items.splice(emptyIndex, 1);
             sticker.renderItem(sticker.stickerEl.children[3]);
           }
+        }
+      }
+
+      // 다른 스티커로의 이동
+      else {
+        const emptyIndex = target.sticker.items.indexOf('EMPTY');
+
+        if (emptyIndex !== -1) {
+          target.sticker.items.splice(emptyIndex, 1);
         } else {
-          if (item.sticker.items.includes('EMPTY')) {
-            item.sticker.items.splice(item.sticker.items.indexOf('EMPTY'), 1);
-          } else {
-            item.sticker.items = item.sticker.items.filter(
-              (deleteItem) =>
-                item.itemEl.id.split('-')[1] !==
-                deleteItem.itemEl.id.split('-')[1]
-            );
-          }
-
-          toSticker.items[insertIndex ?? 0] = item;
-
-          item.sticker.renderItem(item.sticker.stickerEl.children[3]);
-          item.sticker = toSticker;
-          toSticker.renderItem(toSticker.stickerEl.children[3]);
+          target.sticker.items = target.sticker.items.filter(
+            (item) => targetEl.id !== item.itemEl.id
+          );
         }
 
-        draggingItem = null;
+        toSticker.items[insertIndex ?? 0] = target;
 
-        document.removeEventListener('mousemove', onMouseMove);
-        document.removeEventListener('mouseup', onMouseUp);
-      };
+        target.sticker.renderItem(target.sticker.stickerEl.children[3]);
+        toSticker.renderItem(toSticker.stickerEl.children[3]);
+        target.sticker = toSticker;
+      }
 
-      document.addEventListener('mousemove', onMouseMove);
-      document.addEventListener('mouseup', onMouseUp);
-    });
+      draggingItem = null;
+
+      document.removeEventListener('mousemove', onMouseMove);
+      document.removeEventListener('mouseup', onMouseUp);
+    };
+
+    document.addEventListener('mousemove', onMouseMove);
+    document.addEventListener('mouseup', onMouseUp);
   }
 }
